Add files via upload
In this project, we will create a product page design that can be used on an e-commerce website. On this product page design, we will display the product image gallery and product description. We will also add an option to choose the product variation before buying. We will make this e-commerce product page using HTML, CSS, and JavaScript.